create edit student name function

+++ User/Student/student.c
@@ -12,7 +12,7 @@ Return: it returns no thing.                                                    
 Usage: it is used to edit name of the student that we receive its ID.            |
 ----------------------------------------------------------------------------------------------
 */
-void Edit_student_name(s_student *head_student,int ID){
+void Edit_student_name(s_student *head_student,long ID){
 int flag = 0; /* use flag to check id is exist or not flag=1 this means id is exist ,
               if flag=0 this means id isn't exist*/
 char *edit_name ;
